+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.10.0
+* Add a growable `BitList`, make it behave more like `List<bool>`.
+* Optimize various `BitList` operations, i.e. `fill`, `count`, ...
+
 ## 3.9.0
 * Various extension to work with bits.
 * Add iterator over `BitList.indices` that are true or false.